requirements.txt
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.2.17
django-filter==2.4.0
django-rest-swagger==2.2.0
djangorestframework==3.12.2
idna==2.10
importlib-metadata==3.1.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==2.11.2
Markdown==3.3.3
MarkupSafe==1.1.1
openapi-codec==1.3.2
packaging==20.8
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.0
pytz==2020.4
requests==2.25.0
simplejson==3.17.2
sqlparse==0.4.1
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
zipp==3.4.0
